Custom init -> Plugin init: Updated integration test
