--- test/kernel/integration/custom_init/README.md
@@ -1,1 +1,0 @@
-# Test custom pre-service initialization routines
