--- test/kernel/integration/custom_init/vm.json
@@ -1,1 +1,0 @@
-{"image" : "test_custom_init.img" }
