+++ test/kernel/integration/plugin_init/README.md
@@ -1,0 +1,1 @@
+# Test pre-service plugin initialization routines
