+++ test/kernel/integration/plugin_init/test.py
