+++ test/kernel/integration/plugin_init/vm.json
@@ -1,0 +1,1 @@
+{"image" : "test_plugin_init.img" }
